# docs/swagger.yaml
definitions:
  app.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      msg:
        type: string
    type: object
  models.BankCardInfo:
    properties:
      bankBranchName:
        description: 支行名称
        type: string
      bankCard:
        description: 银行卡号
        type: string
      bankName:
        description: 银行名称
        type: string
      password:
        description: 给银行卡信息设置的密码
        type: string
      userId:
        type: integer
      userName:
        description: 开户名
        type: string
    type: object
  models.RequestBankCardInfo:
    properties:
      bankCardInfo:
        $ref: '#/definitions/models.BankCardInfo'
        type: object
      code:
        description: 验证码
        type: string
    type: object
info:
  contact: {}
  description: A web of gin
  license: {}
  title: Golang Gin API
  version: "1.0"
paths:
  /api/v1/agent/bind:
    post:
      parameters:
      - description: user_id
        in: body
        name: user_id
        schema:
          type: integer
      - description: agent_id
        in: body
        name: agent_id
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 绑定上级
      tags:
      - 代理
  /api/v1/bank:
    get:
      consumes:
      - application/json; charset=utf-8
      parameters:
      - description: user_id
        in: body
        name: user_id
        schema:
          type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 获取银行卡信息
      tags:
      - 银行卡
  /api/v1/bink/bind:
    post:
      parameters:
      - description: 请求的json结构
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/models.RequestBankCardInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 绑定银行卡
      tags:
      - 银行卡
...
  /api/v1/order:
    post:
      parameters:
      - description: user_id
        in: body
        name: user_id
        schema:
          type: integer
      - description: 游戏
        in: body
        name: game_type
        schema:
          type: integer
      - description: 订单类型
        in: body
        name: order_type
        schema:
          type: integer
      - description: 代练类型
        in: body
        name: instead_type
        schema:
          type: integer
      - description: 游戏区服
        in: body
        name: game_zone
        schema:
          type: integer
      - description: 铭文等级
        in: body
        name: runes_level
        schema:
          type: integer
      - description: 英雄数量
        in: body
        name: hero_num
        schema:
          type: integer
      - description: 当前段位
        in: body
        name: cur_level
        schema:
          type: integer
      - description: 目标段位
        in: body
        name: target_level
        schema:
          type: integer
      - description: 游戏账号
        in: body
        name: game_account
        schema:
          type: string
      - description: 游戏密码
        in: body
        name: game_pwd
        schema:
          type: string
      - description: 游戏角色名
        in: body
        name: game_role
        schema:
          type: string
      - description: 验证手机
        in: body
        name: game_phone
        schema:
          type: string
      - description: 保证金
        in: body
        name: margin
        schema:
          type: integer
      - description: 有防沉迷
        in: body
        name: anti_addiction
        schema:
          type: integer
      - description: 有指定英雄
        in: body
        name: designate_hero
        schema:
          type: integer
      - description: 指定英雄
        in: body
        name: hero_name
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 下单
      tags:
      - 下单
  /api/v1/order/all:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: Get 获取订单列表
      tags:
      - 接单
  /api/v1/order/confirm:
    post:
      parameters:
      - description: user_id
        in: body
        name: user_id
        schema:
          type: integer
      - description: order_id
        in: body
        name: order_id
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 完成订单
      tags:
      - 订单
  /api/v1/order/finish:
    post:
      parameters:
      - description: user_id
        in: body
        name: user_id
        schema:
          type: integer
      - description: order_id
        in: body
        name: order_id
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 完成订单
      tags:
      - 订单
  /api/v1/order/take:
    post:
      parameters:
      - description: user_id
        in: body
        name: user_id
        schema:
          type: integer
      - description: order_id
        in: body
        name: order_id
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 接单
      tags:
      - 接单
...
swagger: "2.0"